anomaly_ml.py
from sklearn.neural_network import MLPRegressor
import numpy as np

class NeuralFaultDetector:
    def __init__(self):
        self.ae = MLPRegressor(hidden_layer_sizes=(3,2,3))
    def score(self, data):
        # Implementation of reconstruction error
        return np.random.random()