requirements.txt
streamlit
numpy
scikit-learn
paho-mqtt
plotly
scipy
psycopg2-binary